docs(assignment-3): 📝 more docs/comments
Leaves a note deprecating part 3 of the assignment as it was removed
from the assignment in a PR post fork. Describes why it was left up in
the markdown doc as well.

Additionally, update some of the comments+ add docstrings to the
methods explaining the use of dicts instead of a sorted string.